% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_blocos_id.R
\name{extrair_blocos_id}
\alias{extrair_blocos_id}
\title{Extrai os blocos partidários da Câmara dos Deputados por \emph{id}.}
\arguments{
\item{url_base}{Identificador único do órgão.}

\item{id}{A URL do JSON que contém os dados básicos dos blocos partidários, a partir de seu id.}
}
\value{
Um dataframe contendo os dados básicos do bloco partidário.
}
\description{
Esta função consulta a API de Dados Abertos da CD e extrai os dados básicos dos blocos partidários
a partir de um \emph{id} específico.
}
\examples{
df_blocos_id <- extrair_blocos_id(id = 584)
}
