% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_blocos_partidarios.R
\name{extrair_blocos_partidarios}
\alias{extrair_blocos_partidarios}
\title{Extrai os dados básicos dos \emph{blocos partidários} de uma \emph{Legislatura}
específica.}
\arguments{
\item{url_base}{A URL do JSON que contém os dados básicos de blocos partidários.
Nas atividades parlamentares, os partidos podem se juntar em blocos partidários.
Quando associados, os partidos passam a trabalhar como se fossem um "grande partido",
com um só líder e um mesmo conjunto de vice-líderes. Os blocos só podem existir
até o fim da legislatura em que foram criador. Se forem passados números de
legislaturas com o parâmetro idLegislatura, são listados os blocos formados e
extintos nessas legislaturas.}

\item{idLegislatura}{Identificador único do bloco partidário.}

\item{itens}{Número de itens retornado por página.}

\item{ordem}{Ordem de classificação.}

\item{ordenarPor}{Critério de classificação.}
}
\value{
Um dataframe contendo os dados básicos do bloco.
}
\description{
Esta função consulta a API de Dados Abertos da CD e extrai os dados básicos
dos \emph{blocos partidários} de uma \emph{Legislatura} específica.
}
\examples{
extrair_blocos(idLegislatura = 56)
}
