% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_deputados_legislatura.R
\name{extrair_deputados_legislatura}
\alias{extrair_deputados_legislatura}
\title{Extrai os dados básicos dos \emph{deputados federais} de uma dada \emph{Legislatura}.}
\arguments{
\item{url_base}{A base da URL do JSON que contém os dados básicos dos deputados.}

\item{idLegislatura}{Identificador único da legislatura; caso não informado,
retorna os deputados em exercício.}

\item{dataInicio}{Data início; caso não informado, é ignorado.}

\item{dataFim}{Data fim; caso não informado, é ignorado.}

\item{itens}{Número de itens retornados por página; padrão = 1000 (mesmo da API).}

\item{ordem}{Ordem dos dados (crescente ou decrescente).}

\item{ordenarPor}{Critério de ordenação dos dados.}
}
\value{
Dataframe contendo os dados básicos dos deputados.
}
\description{
Esta função consulta a API de Dados Abertos da CD e extrai os dados básicos
dos \emph{deputados federais}, segundo os critérios informados.
Caso apenas o \emph{id} da \emph{Legislatura} seja passado, a função retorna
os dados dos deputados federais desta legislatura e das seguintes, caso ele
tenha sidoreeleito e se encontre em exercício.
Caso, além do \emph{id} da \emph{Legislatura}, a \emph{data inicial} e a
\emph{data final} sejam passadas, a função retornará só os dados dos deputados
que atendam aos três critérios.
}
\examples{
extrair_deputados_legislatura(idLegislatura = 56)
}
