% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_legislatura.R
\name{extrair_legislatura}
\alias{extrair_legislatura}
\title{Extrai os dados básicos de uma \emph{Legislatura} específica.}
\arguments{
\item{url_base}{A base da URL do JSON que contém os dados básicos das
legislaturas.}

\item{id_legislatura}{O identificador da legislatura. Se nenhuma for passada,
retorna os dados da legislatura atual.}
}
\value{
Um dataframe contendo os dados básicos de uma legislatura específica
}
\description{
Esta função consulta a API de Dados Abertos da Câmara dos Deputados e
extrai os dados básicos de uma \emph{Legislatura} específica.
}
\examples{
df_legislatura <- extrair_legislatura()
}
