% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_legislaturas.R
\name{extrair_legislaturas}
\alias{extrair_legislaturas}
\title{Extrai os dados básicos de todas as legislaturas disponíveis}
\arguments{
\item{url_base}{A base da URL do JSON que contém os dados básicos das legislaturas.}

\item{ordem}{Ordem dos dados das legislaturas.}

\item{ordenarPor}{Critério de ordenação dos dados das legislaturas.}
}
\value{
Um dataframe contendo os dados básicos das legislaturas.
}
\description{
Esta função consulta a API de Dados Abertos da CD e extrai os dados básicos de todas as legislaturas disponíveis.
}
\examples{
extrair_legislaturas()
}
\author{
Marcus Chevitarese
}
