% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_lideres_legislatura.R
\name{extrair_lideres_legislatura}
\alias{extrair_lideres_legislatura}
\title{Extrai os dados básicos dos \emph{deputados federais} de uma dada
\emph{legislatura}.}
\arguments{
\item{url_base}{A base da URL do JSON que contém os dados básicos dos deputados.}

\item{idLegislatura}{Identificador único da legislatura; caso não informado,
retorna os deputados em exercício.}

\item{itens}{Número de itens retornados por página; padrão = 15 (mesmo da API).}
}
\value{
Dataframe contendo os dados básicos dos deputados.
}
\description{
Esta função consulta a API de Dados Abertos da Câmara dos Deputados e extrai
os dados básicos dos \emph{deputados federais}, segundo os critérios informados.
Caso apenas o \emph{id} da \emph{Legislatura} seja passado,
a função retorna os dados dos deputados federais desta \emph{Legislatura} e
das seguintes, caso ele tenha sido reeleito e se encontre em exercício.
Caso, além do \emph{id} da \emph{Legislatura}, o data início e
data fim sejam passadas, a função retornará só os dados dos deputados que
atendam aos três critérios.
}
\examples{
df_lideres_legislatura <- extrair_lideres_legislatura()
}
