% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_membros_frentes_parlamentares.R
\name{extrair_membros_frentes}
\alias{extrair_membros_frentes}
\title{Extrai os membros de uma frente parlamentar da Câmara dos Deputados.}
\arguments{
\item{url_base}{A URL do JSON que contém os dados básicos dos membros de uma frente parlamentar.}

\item{id}{Identificador único da uma frente parlamentar; caso não informado, retorna as frentes
de todas as legislaturas.}
}
\value{
Um dataframe contendo os dados básicos da frente.
}
\description{
Esta função consulta a API de Dados Abertos da Câmara dos Deputados e extrai
os dados básicos dos membros de uma \emph{frente parlamentar} de uma
\emph{legislatura} específica.
}
\examples{
df_membros_frentes <- extrair_membros_frentes(id = 55628)
}
