% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_mesa_legislatura.R
\name{extrair_mesa_legislatura}
\alias{extrair_mesa_legislatura}
\title{Extrair os dados dos membros da \emph{Mesa-Diretora} da Câmara dos Deputados,
em uma dada \emph{Legislatura}.}
\arguments{
\item{url_base}{A base da URL do JSON que contém os dados básicos das legislaturas.}

\item{id_legislatura}{Argumento que identifica a Legislatura. Caso não seja
informado, a função retorna os dados da Legislatura atual.}
}
\value{
Um dataframe contendo os dados dos membros da Legislatura.
}
\description{
Esta função consulta a API de Dados Abertos da Câmara dos Deputados e extrai
os dados básicos dos membros da \emph{Mesa-Diretora}, em uma dada
\emph{Legislatura}.
}
\examples{
df_mesa <- extrair_mesa_legislatura()
}
