% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_partidos_membros.R
\name{extrair_partidos_membros}
\alias{extrair_partidos_membros}
\title{Extrai os dados dos \emph{membros} de um dado \emph{partido político} na
Câmara dos Deputados.}
\arguments{
\item{url_base}{A base da URL do JSON que contém os dados básicos dos membros de um partido.}

\item{id}{Argumento que identifica um partido político.}
}
\value{
Um dataframe contendo os dados dos membros do partido político.
}
\description{
Esta função consulta a API de Dados Abertos da CD e extrai os dados dos
\emph{membros} de um dado \emph{partido político}, a partir de seu \emph{id}.
}
\details{
Usar em conjunto com a função \emph{extrair_partidos()}, para obter o
\emph{id} do \emph{partido}.
}
\examples{
df_partido_membros <- extrair_partido_membros(id = 37906)
}
