% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_proposicoes_autores.R
\name{extrair_proposicoes_autores}
\alias{extrair_proposicoes_autores}
\title{Extrai dados dos autores de \emph{proposições legislativas} da Câmara dos
Deputados, por meio do \emph{id} da proposição.}
\arguments{
\item{url_base}{A base da URL do JSON que contém os dados básicos das
proposições.}

\item{id}{Identificadores de uma ou mais proposições, separados por vírgulas.}
}
\value{
Um dataframe contendo os dados das proposições ou uma mensagem de
erro se a consulta falhar.
}
\description{
Esta função consulta a API de Dados Abertos da Câmara dos Deputados e retorna
dados de \emph{autoria} das proposições, a partir do \emph{id} da proposição
legislativa.
}
\examples{
df_proposicoes_autores <- extrair_proposicoes_autores(id = "18156")
}
