% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_proposicoes_id.R
\name{extrair_proposicoes_id}
\alias{extrair_proposicoes_id}
\title{Extrai dados de \emph{proposições legislativas} da Câmara dos Deputados,
por meio de seu \emph{id}.}
\arguments{
\item{url_base}{A base da URL do JSON que contém os dados da proposição.}

\item{id}{Identificador da proposição legislativa.}
}
\value{
Um dataframe contendo os dados da proposição legislativa ou uma
mensagem de erro, se a consulta falhar.
}
\description{
Esta função consulta a API de Dados Abertos da Câmara dos Deputados e retorna
dados detalhados de uma \emph{proposição legislativa}, a partir de seu
\emph{id}.
}
\examples{
df_proposicoes_id <- extrair_proposicoes_id(id = "18157")
}
