% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_proposicoes_relacionadas.R
\name{extrair_proposicoes_relacionadas}
\alias{extrair_proposicoes_relacionadas}
\title{Extrai dados das \emph{proposições legislativas} da Câmara dos Deputados
associadas a uma dada proposição.}
\arguments{
\item{url_base}{A base da URL do JSON que contém os dados das proposições.}

\item{id}{Identificadores de uma ou mais proposições, separados por vírgulas.}
}
\value{
Um dataframe contendo os dados das proposições ou uma mensagem de
erro se a consulta falhar.
}
\description{
Esta função consulta a API de Dados Abertos da Câmara dos Deputados e retorna
uma lista de informações básicas sobre proposições que de alguma forma se
relacionam com a proposição identificada por {id} (pareceres, requerimentos,
substitutivos, etc.).
}
\examples{
df_proposicoes_relacionadas <- extrair_proposicoes_relacionadas(id = "18156")
}
