% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_proposicoes_temas.R
\name{extrair_proposicoes_temas}
\alias{extrair_proposicoes_temas}
\title{Extrai dados dos temas das \emph{proposições legislativas} da Câmara dos
Deputados pelo {id} da \emph{proposição}.}
\arguments{
\item{url_base}{A base da URL do JSON que contém os dados da proposição.}

\item{id}{Identificadores de uma ou mais proposições, separados por vírgulas.}
}
\value{
Um dataframe contendo os dados das proposições ou uma mensagem de
erro se a consulta falhar.
}
\description{
Esta função consulta a API de Dados Abertos da Câmara dos Deputados e
retorna dados da(s) área(s) temática(s) à(s) qual(is) a proposição
identificada por {id} se relaciona (segundo classificação do CEDI).
}
\examples{
df_proposicoes_temas <- extrair_proposicoes_temas(id = "18156")
}
