% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_proposicoes_tramitacoes.R
\name{extrair_proposicoes_tramitacoes}
\alias{extrair_proposicoes_tramitacoes}
\title{Extrai dados do histórico de \emph{tramitação} de
\emph{proposições legislativas} da Câmara dos Deputados pelo {id} da
proposição.}
\arguments{
\item{url_base}{A base da URL do JSON que contém os dados básicos das proposições.}

\item{id}{Identificadores de uma ou mais proposições, separados por vírgulas.}
}
\value{
Um dataframe contendo os dados das proposições ou uma mensagem de erro se a consulta falhar.
}
\description{
Esta função consulta a API de Dados Abertos da Câmara dos Deputados e
retorna dados sobre um estágio da tramitação da proposição, obtido por meio
do seu \emph{id}.
Esta representação das tramitações é provisória.
}
\examples{
df_proposicoes_tramitacoes <- extrair_proposicoes_tramitacoes(id = "18156")
}
