% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_proposicoes_votacoes.R
\name{extrair_proposicoes_votacoes}
\alias{extrair_proposicoes_votacoes}
\title{Extrai dados de histórico de tramitação de votações de \emph{proposições}
da Câmara dos Deputados pelo \emph{id} da proposição.}
\arguments{
\item{url_base}{A base da URL do JSON que contém os dados básicos das proposições.}

\item{id}{Identificadores de uma ou mais proposições, separados por vírgulas.}
}
\value{
Um dataframe contendo os dados das proposições ou uma mensagem de
erro se a consulta falhar.
}
\description{
Esta função consulta a API de Dados Abertos da Câmara dos Deputados e
retorna dados sobre uma votação da proposição, obtido por meio do seu {id}.
A proposição pode ter sido objeto da votação ou sido afetado
por esta (resultado da votação). Dados complementares sobre cada votação
listada podem ser obtidos no recurso /votacoes/{id}.
}
\examples{
df_proposicoes_votacoes <- extrair_proposicoes_votacoes(id = "18156")
}
