% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_votacoes.R
\name{extrair_votacoes}
\alias{extrair_votacoes}
\title{Extrai informações básicas sobre as votações ocorridas em eventos nos órgãos
da Câmara dos Deputados.}
\arguments{
\item{url_base}{A base da URL do JSON que contém os dados das votações.}

\item{id}{Identificador único da votação}

\item{idProposicao}{Identificador único da proposição.}

\item{idOrgao}{Identificador único do órgão.}

\item{dataInicio}{Data início. Caso não informado o intervalo de pesquisa,
são retornados dados de todas as votações ocorridas nos últimos 30 dias,
em eventos de todos os órgãos.
Quando informado intervalo, é necessário que as datas sejam de um mesmo ano.
Quando apenas uma delas está presente, são retornadas somente as votações
ocorridas no mesmo ano, antes de dataFim ou após dataInicio.}

\item{dataFim}{Data fim.}

\item{ordenarPor}{Critério de ordenação dos dados.}

\item{ordem}{Ordem dos dados (crescente ou decrescente).}

\item{itens}{Número de itens retornados por página; padrão = 1000 (mesmo da API).}
}
\value{
Dataframe contendo os dados básicos dos membros do partido.
}
\description{
Esta função consulta a API de Dados Abertos da CD e extrai os dados básicos
das votações ocorridas em órgãos da Câmara dos Deputados.
}
\examples{
df_votacoes <- extrair_votacoes(dataInicio = "2024-06-30")
}
