% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_votacoes_orientacoes.R
\name{extrair_votacoes_orientacoes}
\alias{extrair_votacoes_orientacoes}
\title{Extrai os dados das votações segundo as orientações dos partidos políticos,
blocos parlamentares, líder do Governo, Oposição, Maioria e Minoria.}
\arguments{
\item{url_base}{A base da URL do JSON que contém os dados das votações.}

\item{id}{Identificador único da votação.}
}
\value{
Dataframe contendo os dados básicos das votações.
}
\description{
Uma liderança também pode orientar para liberar a bancada (para que cada
deputado vote como quiser), ou entrar em obstrução (parlamentares não são
contados para o quórum de votação), ou ainda podem votar sim ou não.
Atualmente, só existem dados de orientação de votação para as votações em
\emph{Plenário}.
}
\details{
Esta função consulta a API de Dados Abertos da CD e extrai os dados das
orientações dos partidos políticos, blocos e bancadas de Governo, Oposição,
Maioria e Minoria.
}
\examples{
df_votacoes_orientacoes <- extrair_votacoes_orientacoes(id = "994689-78")
}
