% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_votacoes_votos.R
\name{extrair_votacoes_votos}
\alias{extrair_votacoes_votos}
\title{Extrai informações básicas sobre como cada parlamentar votou em uma votação
nominal e aberta.}
\arguments{
\item{url_base}{A base da URL do JSON que contém os dados das votações.}

\item{id}{Identificador único da votação.}
}
\value{
Dataframe contendo os dados básicos das votações.
}
\description{
Esta função consulta a API de Dados Abertos da CD e extrai os dados básicos
dos votos ocorridos nas votações em órgãos da Câmara dos Deputados, por meio
do \emph{id} da \emph{votação}.
}
\examples{
df_votacoes_votos <- extrair_votacoes_votos(id = "994689-78")
}
